requirements.txt
requests
Pillow
piexif
